chore: creating separate plugin for block-quote
