+++ packages/react-tinacms-editor/src/plugins/Blockquote/commands/index.test.ts
@@ -16,9 +16,9 @@ limitations under the License.
 
 */
 
-import { PMTestHarness } from '../prosemirror-test-utils'
-import { defaultSchema } from '../schema/default-schema'
-import { liftBlockquote } from './blockquote-commands'
+import { PMTestHarness } from '../../../prosemirror-test-utils'
+import { defaultSchema } from '../../../schema/default-schema'
+import { liftBlockquote } from '.'
 
 const {
   forDoc,
