+++ packages/react-tinacms-editor/src/plugins/Image/ToolbarComponent/index.tsx
@@ -21,7 +21,7 @@ import styled, { css } from 'styled-components'
 import { Button } from '@tinacms/styles'
 import { Input } from '@tinacms/fields'
 import { MediaIcon, UploadIcon, CloseIcon } from '@tinacms/icons'
-import { insertImage } from '../../../commands/image-commands'
+import { insertImage } from '../commands'
 import { MenuButton, MenuButtonDropdown } from '../../Menu/MenuComponents'
 import { Dismissible } from 'react-dismissible'
 import { useEditorStateContext } from '../../../context/editorState'
