+++ packages/react-tinacms-editor/src/plugins/Image/plugin.ts
@@ -25,7 +25,7 @@ import {
 } from 'prosemirror-state'
 import { EditorView, DecorationSet, Decoration } from 'prosemirror-view'
 import { Slice } from 'prosemirror-model'
-import { insertImageList } from '../../commands'
+import { insertImageList } from './commands'
 import { ImageView } from './nodeView'
 
 export const imagePluginKey = new PluginKey('image')
