+++ packages/react-tinacms-editor/src/plugins/keymap.ts
@@ -34,9 +34,9 @@ import {
 import { EditorView } from 'prosemirror-view'
 import { findParentNodeOfType } from 'prosemirror-utils'
 import * as TableCommands from 'prosemirror-tables'
-import { liftBlockquote } from '../commands/blockquote-commands'
+import { liftBlockquote } from './Blockquote/commands'
 import { toggleBulletList, toggleOrderedList } from '../commands/list-commands'
-import { deleteEmptyHeading, toggleHeader } from '../commands/heading-commands'
+import { deleteEmptyHeading, toggleHeader } from './Block/commands'
 import { undo, redo } from 'prosemirror-history'
 import { undoInputRule } from 'prosemirror-inputrules'
 import { KeymapPlugin } from '../types'
